Gulpfile builds all dependencies into .generated
Previously the Gulpfile would not build all of the vendor assets
into the .generated folder, and the webserver would then serve vendor
assets directly from bower_components. We now yank the vendor assets
from bower_components and place them in .generated/vendor so that
the .generated dir can be served on the localhost or heroku via the node
webserver, or can be rsynced to be served by a static server like Apache
or served via Github Pages.